LPS-53199 Move getStrutsAction() to a base class because it's not wiki specific
